refactor: use format interface for both v0 and v1 lightGodwoken
