+++ src/light-godwoken/index.ts
@@ -1,4 +1,5 @@
-import LightGodwoken from "./lightGodwoken";
 import LightGodwokenProvider from "./lightGodwokenProvider";
+import LightGodwokenV0 from "./LightGodwokenV0";
+import LightGodwokenV1 from "./LightGodwokenV1";
 
-export { LightGodwokenProvider, LightGodwoken };
+export { LightGodwokenProvider, LightGodwokenV1, LightGodwokenV0 as LightGodwoken };
