+++ src/v1.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import DefaultLightGodwoken from "./light-godwoken/lightGodwoken";
+import { LightGodwokenV1 } from "./light-godwoken/index";
 import DefaultLightGodwokenProvider from "./light-godwoken/lightGodwokenProvider";
 import detectEthereumProvider from "@metamask/detect-provider";
 import { LightGodwoken } from "./light-godwoken";
@@ -12,7 +12,7 @@ const Page: React.FunctionComponent = () => {
 
         if (!accounts || !accounts[0]) return;
 
-        const instance = new DefaultLightGodwoken(new DefaultLightGodwokenProvider(accounts[0], ethereum));
+        const instance = new LightGodwokenV1(new DefaultLightGodwokenProvider(accounts[0], ethereum));
         setLightGodwoken(instance);
       });
 
@@ -21,14 +21,14 @@ const Page: React.FunctionComponent = () => {
 
         if (!accounts || !accounts[0]) return setLightGodwoken(undefined);
 
-        const provider = new DefaultLightGodwoken(new DefaultLightGodwokenProvider(ethereum.selectedAddress, ethereum));
+        const provider = new LightGodwokenV1(new DefaultLightGodwokenProvider(ethereum.selectedAddress, ethereum));
         setLightGodwoken(provider);
       });
     });
   }, []);
   const handleClick = () => {
     console.log("handle click", lightGodwoken);
-    lightGodwoken!.withdrawV1WithEvent({
+    lightGodwoken!.withdrawWithEvent({
       capacity: "0x9c7652400", //420
       amount: "0x0",
       sudt_script_hash: `0x${"00".repeat(32)}`,
